==> notebooks/pdui/3_mortalidade.ipynb <==
# Cálculo de taxa de mortalidade

## Preparação do Notebook

Na célula adiante são feitos procedimentos de preparação do notebook para trabalho com os dados, tais como:
1. mapeamento do diretório em que se encontra o pacote br_demography;
2. importação de bibliotecas necessárias para processamento dos dados;
3. carregametno de project_id para faturamento no Google Cloud, conforme requerido pela biblioteca basedosdados.

# adicionando br_demography to Python Path
import sys 
sys.path.append("../../")

# carregamento de bibliotecas para processamento dos dados
from br_demography  import municipality_deaths as md
import basedosdados as bd
from dotenv import load_dotenv
import os
import pandas as pd # importação de biblioteca de manipulação de dados tabulares
import matplotlib.pyplot as plt # importação de biblioteca de plotagem de gráficos 
import numpy as np # importação de biblioteca para cálculos matemáticos
from scipy.optimize import curve_fit # importação de função a ser utilizada para ajuste de modelo preditivo 
from itertools import product # importação de função voltada para otimização de laços de iteração
import warnings
from itertools import product



# Carregando as variáveis de ambiente do arquivo .env
load_dotenv(dotenv_path='../../.env')

# Carregando o project_id para faturamento no Google Cloud
project_id = os.getenv('GOOGLE_CLOUD_PROJECT_ID')

## Carregamento, tratamento e salvamento de microdados sobre nascidos vivos

Com base nos dados do Sistema de Informações sobre Mortalidade - SIM, sao realizados os seguintes processamentos:
1. carregamento do arquivo com os municípios de interesse;
2. _loop_ para cada município de interesse;
3. download dos microdados SIM por consulta SQL à _DataLake_ da Base dos Dados;
4. tratamento dos dados (padronização de faixas etárias decenais, indexação, inserção de valor "0" para coortes com valor nulo, padronização de tipagem, adequação de colunas, etc.);
5. concatenação dos dados de cada município no DataFrae df_deaths
6. salvamento dos no arquivo obitos_rmc_2000_2022.csv

df_mun = pd.read_csv('../../br_demography/source/tab/cod_mun.csv', sep=';', index_col='mun_id')

dict_dfs = dict()

for mun_id, mun in df_mun.itertuples():
    df_deaths_mun = md.query_deaths(mun_id=mun_id, project_id=project_id, start_year=2000, end_year=2022)
    df_deaths_mun = md.standard_age_groups(df=df_deaths_mun, age_group_csv_path='../../br_demography/source/tab/faixas_etarias_censo_2000_2010.csv')
    df_deaths_mun['Município'] = mun
    df_deaths_mun = df_deaths_mun.reset_index().set_index(['Município', 'Sexo','Faixa Etária'])
    
    dict_dfs[mun] = df_deaths_mun

df_deaths = pd.concat(objs=dict_dfs.values())

df_deaths.to_csv(path_or_buf='../../br_demography/results/tab/obitos_rmc_2000_2022.csv', sep=';', decimal=',', encoding='utf-16')

### Visualização parcial do DataFrame df_deaths, com microdados de óbitos por coorte

df_deaths.iloc[np.r_[0:2, -2:0]]

### Carregamento dos dados de população entre 2000 e 2022

df_pop = pd.read_csv(filepath_or_buffer='../../br_demography/results/tab/pop_municipios_rmc_2000_2022_estimativa_intercensitaria.csv', sep=';')
#df_pop[df_pop['Faixa Etária'].isin(values=df_deaths.index)]
df_pop.set_index(keys=['Município', 'Sexo','Faixa Etária'], inplace=True)
df_pop.columns = df_pop.columns.astype(int)
df_pop.iloc[np.r_[0:2, -2:0]]

## Cálculo de Taxa de Mortalidade

# criação de dataframe com taxa de mortalidade a cada mil habitantes
df_mortalidade = (df_deaths / df_pop).dropna(axis=0) * 1000

df_mortalidade.drop(columns=[2020, 2021, 2022], inplace=True)


## Modelagem das tendências

1. definição de modelos (linear e exponencial)
4. para coortes com tendência geral de crescimento, aplicação de modelo linear;
5. para coortes com tendência geral de decréscimo, modelo exponencial;
6. resultados são armazenados em DataFrame; 
6. plotagem dos resultados;  
6. resultados são armazenados no arquivo taxa_natalidade_projetada_rmc_2022_2045_2.csv.

from scipy.stats import iqr

# Função para o modelo exponencial decrescente
def modelo_exponencial(x, constante, coeficiente):
    return constante * np.exp(coeficiente * (-x))

# Função para o modelo Linear - casos de crescimento
def modelo_linear(x, constante, coeficiente):
    return coeficiente * x + constante

# DataFrame de resultados
resultados = pd.DataFrame(columns=['Município', 'Sexo', 'Faixa Etária','Tipo de Modelo','Constante', 'Coeficiente'])

# Combinações de sexo e faixa etária
df_mortalidade.reset_index(inplace=True)

plt.ioff()

for mun_id, mun in df_mun.itertuples():
    print(mun)

    sexos = df_mortalidade[df_mortalidade['Município'] == mun]['Sexo'].unique()
    faixas_etarias = df_mortalidade[df_mortalidade['Município'] == mun]['Faixa Etária'].unique()

    num_plots = len(sexos) * len(faixas_etarias)
    num_colunas = 3
    num_linhas = (num_plots + num_colunas - 1) // num_colunas

    # Criar subplots para cada combinação de sexo e faixa etária
    fig, axs = plt.subplots(num_linhas, num_colunas, figsize=(15, 18))
    fig.subplots_adjust(wspace=0.5, hspace=0.5)

    for i, (sexo, faixa_etaria) in enumerate(product(sexos, faixas_etarias)):
        if i >= num_plots:
            break  # Sair se todos os subplots já estiverem preenchidos
        
        # Filtrar o DataFrame para a combinação atual
        df_filtro = df_mortalidade[(df_mortalidade['Município'] == mun) & (df_mortalidade['Sexo'] == sexo) & (df_mortalidade['Faixa Etária'] == faixa_etaria)]
        
        # Extrair os dados de anos e mortalidade
        anos = np.array(df_filtro.columns[4:], dtype=int) - 2000
        dados_mortalidade = df_filtro.iloc[0, 4:].values.astype(float)
        anos_todo_intervalo = np.arange(0, 47)

        # Filtrar anos com dados de mortalidade não-zero
        anos_com_dados = anos[dados_mortalidade > 0]
        dados_mortalidade = dados_mortalidade[dados_mortalidade > 0]

        outlier_factor = 1.2

        IQR = iqr(dados_mortalidade)
        lower_bound = np.percentile(dados_mortalidade, 25) - outlier_factor * IQR
        upper_bound = np.percentile(dados_mortalidade, 75) + outlier_factor * IQR

        non_outliers_mask = (dados_mortalidade >= lower_bound) & (dados_mortalidade <= upper_bound)
        anos_com_dados = anos_com_dados[non_outliers_mask]
        dados_mortalidade = dados_mortalidade[non_outliers_mask]


        # Verificar se a tendência é de crescimento ou decrescimento
        coef_angular = np.polyfit(anos_com_dados, dados_mortalidade, 1)[0]
        if coef_angular >= 0:

            # Ajustar o modelo linear aos dados
            try:
                params, _ = curve_fit(modelo_linear, anos_com_dados, dados_mortalidade)
                constante, coeficiente = params
                previsao = modelo_linear(anos_todo_intervalo, constante, coeficiente)
                tipo_modelo = 'Linear'

            except RuntimeError:
                constante, coeficiente = np.nan, np.nan
                previsao = np.nan * anos
        else:
            # Ajustar o modelo exponencial aos dados
            try:
                params, _ = curve_fit(modelo_exponencial, anos_com_dados, dados_mortalidade, maxfev=10000)
                constante, coeficiente = params
                previsao = modelo_exponencial(anos_todo_intervalo, constante, coeficiente)
                tipo_modelo = 'Exponencial'

            except RuntimeError:
                constante, coeficiente = np.nan, np.nan
                previsao = np.nan * anos



        # Criar um DataFrame temporário com os resultados atuais
        df_temp = pd.DataFrame({'Município':[mun],'Sexo': [sexo], 'Faixa Etária': [faixa_etaria],
                                'Tipo de Modelo':[tipo_modelo],'Constante': [constante], 'Coeficiente': [coeficiente]})
        
        # Concatenar o DataFrame temporário ao DataFrame de resultados
        resultados = pd.concat([resultados, df_temp], ignore_index=True)

        
        # Plotar os dados originais e a curva ajustada
        row, col = divmod(i, num_colunas)
        axs[row, col].scatter(anos_com_dados + 2000, dados_mortalidade, label='Dados de mortalidade', marker='o')
        axs[row, col].plot(anos_todo_intervalo + 2000, previsao, label=f'Modelo {tipo_modelo} Ajustado', linestyle='--', color='red')
        axs[row, col].set_xlabel('Ano')
        axs[row, col].set_ylabel('Taxa de Óbitos')
        axs[row, col].set_title(f'Sexo: {sexo} | Faixa Etária: {faixa_etaria}')
        axs[row, col].grid(True)
        axs[row, col].legend()

    # Add a single legend below all the plots in the right corner
#    plt.legend(bbox_to_anchor=(.4, -.7), loc='lower left', borderaxespad=0.)


    fig.suptitle(f'Projeção de Mortalidade por Sexo e Faixa - {mun} \n', fontsize=15, fontweight='bold')
    plt.tight_layout()
#    plt.savefig(f'../../br_demography/results/fig/death_rate/teste/mortalidade_{mun}_non_outliers.svg', format="svg")
    plt.savefig(f'../../br_demography/results/fig/death_rate/teste/mortalidade_{mun}_non_outliers_1dot2IQR.png', dpi=300, bbox_inches='tight')
    plt.close(fig)




import warnings
warnings.filterwarnings('ignore')
warnings.filterwarnings("ignore", category=RuntimeWarning)


resultados.iloc[np.r_[0:2, -2:0]]

### Cálculo da Taxa de Sobrevivência entre 2000 e 2045

new_columns = [*resultados.columns.tolist()]
new_columns.extend(range(2022, 2046))
resultados = resultados.reindex(columns=new_columns, fill_value=0)

for year in resultados.columns[6:].astype(int):
    resultados.loc[:,year] = year-2000

resultados_exp = resultados[resultados['Tipo de Modelo'] == 'Exponencial']
resultados_lin = resultados[resultados['Tipo de Modelo'] == 'Linear']

res = pd.concat(
    objs=[
        1 - modelo_exponencial(resultados_exp.iloc[:, 6:], resultados_exp.Constante.values.reshape((-1,1)), resultados_exp.Coeficiente.values.reshape((-1,1)))/1000
        , 1 - modelo_linear(resultados_lin.iloc[:, 6:], resultados_lin.Constante.values.reshape((-1,1)), resultados_lin.Coeficiente.values.reshape((-1,1)))/1000
        ]
    , axis=0
    
)

resultados = pd.concat(
    objs=[
        resultados_exp.iloc[:,:6]
        , resultados_lin.iloc[:,:6]
        ]
    , axis=0
    
)

resultados = resultados.join(res)

resultados.to_csv(path_or_buf='../../br_demography/results/tab/taxa_sobrevivencia_projetada_pdui_2022_2045.csv', sep=';', decimal=',', encoding='utf-16', index=False)

resultados.iloc[np.r_[0:2, -2:0]]
