==> notebooks/pdui/4_saldo_migratorio.ipynb <==
import sys 
sys.path.append("../../")

from br_demography  import municipality_migration as mm
import basedosdados as bd
from dotenv import load_dotenv
import os
import pandas as pd
import numpy as np

# Carrega as variáveis de ambiente do arquivo .env
load_dotenv(dotenv_path='../../.env')

# Use as variáveis de ambiente conforme necessário
project_id = os.getenv('GOOGLE_CLOUD_PROJECT_ID')

##### AQUI NÃO É POSSÍVEL FAZER UMA AVALIAÇÃO DE QUANTO SERÁ O SALDO MIGRATÓRIO DA REGIÃO, POIS OS DADOS AGREGADOS PARA CADA MUNICÍPIO
##### OU SEJA, O SALDO MIGRATÓRIO DE UM MUNICÍPIO INCLUI OS TROCAS INCLUSIVE COM OUTROS MUNICÍPIOS DA RM, OU SEJA, QUE NAO ALTERAM A POPULAÇÃO REGIONAL

mm.query_total_population(mun_id=4125506, project_id=project_id)

df_mun = pd.read_csv('../../br_demography/source/tab/cod_mun.csv', sep=';', index_col='mun_id')

dict_dfs = dict()

for mun_id, mun in df_mun.itertuples():
    df_immigration_mun_2000 = mm.query_immigration_by_sex_age_2000(mun_id=mun_id, project_id=project_id)
    df_immigration_mun_2000 = mm.standard_age_groups(df=df_immigration_mun_2000, age_group_csv_path='../../br_demography/source/tab/faixas_etarias_censo_2000_2010.csv')
    df_immigration_mun_2000['Município'] = mun
    df_immigration_mun_2000['Ano'] = 2000
    df_immigration_mun_2000 = df_immigration_mun_2000.reset_index().set_index(['Ano', 'Município', 'Sexo','Faixa Etária'])

    df_immigration_mun_2010 = mm.query_immigration_by_sex_age(mun_id=mun_id, project_id=project_id)
    df_immigration_mun_2010 = mm.standard_age_groups(df=df_immigration_mun_2010, age_group_csv_path='../../br_demography/source/tab/faixas_etarias_censo_2000_2010.csv')
    df_immigration_mun_2010['Município'] = mun
    df_immigration_mun_2010['Ano'] = 2010
    df_immigration_mun_2010 = df_immigration_mun_2010.reset_index().set_index(['Ano', 'Município', 'Sexo','Faixa Etária'])

    df_immigration_mun = pd.concat(objs=[df_immigration_mun_2000, df_immigration_mun_2010])

    dict_dfs[mun] = df_immigration_mun

df_immigrantion = pd.concat(objs=dict_dfs.values())
df_immigrantion

dict_dfs = dict()

for mun_id, mun in df_mun.itertuples():

    df_emigration_mun_2000 = mm.query_emigration_by_sex_age_2000(mun_id=mun_id, project_id=project_id)
    df_emigration_mun_2000 = mm.standard_age_groups(df=df_emigration_mun_2000, age_group_csv_path='../../br_demography/source/tab/faixas_etarias_censo_2000_2010.csv')
    df_emigration_mun_2000['Município'] = mun
    df_emigration_mun_2000['Ano'] = 2000
    df_emigration_mun_2000 = df_emigration_mun_2000.reset_index().set_index(['Ano', 'Município', 'Sexo','Faixa Etária'])

    df_emigration_mun_2010 = mm.query_emigration_by_sex_age(mun_id=mun_id, project_id=project_id)
    df_emigration_mun_2010 = mm.standard_age_groups(df=df_emigration_mun_2010, age_group_csv_path='../../br_demography/source/tab/faixas_etarias_censo_2000_2010.csv')
    df_emigration_mun_2010['Município'] = mun
    df_emigration_mun_2010['Ano'] = 2010
    df_emigration_mun_2010 = df_emigration_mun_2010.reset_index().set_index(['Ano', 'Município', 'Sexo','Faixa Etária'])

    df_emigration_mun = pd.concat(objs=[df_emigration_mun_2000, df_emigration_mun_2010])

    dict_dfs[mun] = df_emigration_mun

df_emigrantion = pd.concat(objs=dict_dfs.values())
df_emigrantion

## Calculando a Proporção de cada Faixa Etária no Saldo Migratório

df__proporcao_saldo = ((df_immigrantion - df_emigrantion) /5).round().astype(int)
df__proporcao_saldo = (
    df__proporcao_saldo
    .reset_index()
    .groupby(by=['Ano', 'Município', 'Sexo', 'Faixa Etária'], as_index=False)
    .sum()
    .rename(columns={'Peso':'Saldo Migratório'})
#    .pivot(index=['Ano','Município'], columns=['Sexo', 'Faixa Etária'])
)

# Agrupar por Ano, Município e Sexo e calcular a soma total do saldo migratório para cada grupo
df__proporcao_saldo['Total Saldo Migratório por Grupo'] = df__proporcao_saldo.groupby(['Ano', 'Município', 'Sexo'])['Saldo Migratório'].transform('sum')

# Calcular a proporção do saldo migratório para cada linha em relação ao total do grupo
df__proporcao_saldo['Proporção do Saldo Migratório'] = df__proporcao_saldo['Saldo Migratório'] / df__proporcao_saldo['Total Saldo Migratório por Grupo']

df__proporcao_saldo.to_csv('../../br_demography/results/tab/proporcoes_saldo_migratorio.csv', sep=';', decimal=',', encoding='utf-16', index=False)

df__proporcao_saldo.head()

df_saldo = ((df_immigrantion - df_emigrantion) /5).round().astype(int)
df_saldo = (
    df_saldo
    .reset_index()
    .groupby(by=['Ano', 'Município'], as_index=False)
    .sum()
    .rename(columns={'Peso':'Saldo Migratório'})
    .pivot(index='Município', columns='Ano')
)

df_saldo.head()

df_pop = pd.read_csv(filepath_or_buffer='../../br_demography/results/tab/pop_municipios_rmc_2000_2022_estimativa_intercensitaria.csv', sep=';')
#df_pop[df_pop['Faixa Etária'].isin(values=df_deaths.index)]
df_pop.set_index(keys=['Município', 'Sexo','Faixa Etária'], inplace=True)
df_pop.columns = df_pop.columns.astype(int)
df_pop = (
    df_pop
    .reset_index()
    .groupby(by=['Município'])[2010, 2022]
    .sum()
)

df_pop['Crescimento Pop 2010-2022'] = df_pop[2022] - df_pop[2010]

df_pop.iloc[np.r_[0:2, -2:0]]

df_deaths = pd.read_csv('../../br_demography/results/tab/obitos_rmc_2000_2022.csv', sep=';', decimal=',', encoding='utf-16')

df_deaths = (
    df_deaths
    .groupby(by='Município')
    .sum()
    .loc[:,list(str(year) for year in range(2010, 2022))]
    .sum(axis=1)
    .to_frame()
    .rename(columns={0:'Óbitos 2010 a 2022'})
)

df_births = pd.read_csv('../../br_demography/results/tab/nascimentos_rmc_2000_2022.csv', sep=';', decimal=',', encoding='utf-16')
df_births = (
    df_births
    .drop(columns=['Unnamed: 0'])
    .groupby(by='Município')
    .sum()
    .loc[:,list(str(year) for year in range(2010, 2022))]
    .sum(axis=1)
    .to_frame()
    .rename(columns={0:'Nascimentos 2010 a 2022'})
)

df_pop = pd.concat(
    objs=[
        df_pop
        ,df_births
        ,df_deaths
    ]
    , axis=1
)

df_pop['Crescimento Vegetativo 2010 a 2022'] = df_pop['Nascimentos 2010 a 2022'] - df_pop['Óbitos 2010 a 2022']

df_saldo['2022'] = ((df_pop['Crescimento Pop 2010-2022'] - df_pop['Crescimento Vegetativo 2010 a 2022']) / 12).round(0)

df_saldo = df_saldo.astype(int)
df_saldo.to_csv('../../br_demography/results/tab/saldo_migratorio_anual.csv', sep=';', decimal=',', encoding='utf-16')

(
    df_pop
    .reset_index()
    .groupby(by=['Município'])[[2022]]
    .sum()
)

## Versão Original

df_emigration_2010 = mm.query_emigration_by_sex_age(mun_id=4125506, project_id=project_id)
df_emigration_2010 = mm.standard_age_groups(df=df_emigration_2010, age_group_csv_path='../../br_demography/source/tab/faixas_etarias_censo_2000_2010.csv')
print(df_emigration_2010.sum())
df_emigration_2010

df_immigration_2010 = mm.query_immigration_by_sex_age(mun_id=4125506, project_id=project_id)
df_immigration_2010 = mm.standard_age_groups(df=df_immigration_2010, age_group_csv_path='../../br_demography/source/tab/faixas_etarias_censo_2000_2010.csv')
print(df_immigration_2010.sum())
df_immigration_2010

df_immigration_2000 = mm.query_immigration_by_sex_age_2000(mun_id=4125506, project_id=project_id)
df_immigration_2000 = mm.standard_age_groups(df=df_immigration_2000, age_group_csv_path='../../br_demography/source/tab/faixas_etarias_censo_2000_2010.csv')
print(df_immigration_2000.sum())
df_immigration_2000

df_emigration_2000 = mm.query_emigration_by_sex_age_2000(mun_id=4125506, project_id=project_id)
df_emigration_2000 = mm.standard_age_groups(df=df_emigration_2000, age_group_csv_path='../../br_demography/source/tab/faixas_etarias_censo_2000_2010.csv')
print(df_emigration_2000.sum())
df_emigration_2000

df_saldo_2000 = ((df_immigration_2000 - df_emigration_2000) /5).round().astype(int)
print(df_saldo_2000.sum())
df_saldo_2000

df_saldo_2010 = ((df_immigration_2010 - df_emigration_2010) /5).round().astype(int)
print(df_saldo_2010.sum())
df_saldo_2010

df_saldo.to_csv(path_or_buf='./br_demography/results/tab/saldo_migratorio_anual_2000_2010_campina_2022_2044.csv', sep=';', decimal=',', encoding='utf-16')
