==> notebooks/pdui/5_projecao_populacional.ipynb <==
# Projeção Populacional

## Preparação do Notebook

Na célula adiante são feitos procedimentos de preparação do notebook para trabalho com os dados, tais como:
1. mapeamento do diretório em que se encontra o pacote br_demography;
2. importação de bibliotecas necessárias para processamento dos dados;
3. carregametno de project_id para faturamento no Google Cloud, conforme requerido pela biblioteca basedosdados.

# adding br_demography to Python Path
import sys 
sys.path.append("../../")

from br_demography  import municipality_migration as mm
from br_demography  import municipality_births as mb
from br_demography  import municipality_pop_pyramid as mep

import basedosdados as bd
from dotenv import load_dotenv
import os
import pandas as pd
import numpy as np
# Carrega as variáveis de ambiente do arquivo .env
load_dotenv(dotenv_path='../../.env')

# Use as variáveis de ambiente conforme necessário
project_id = os.getenv('GOOGLE_CLOUD_PROJECT_ID')

# Projeção para 2032

## População do Ano Inicial

categories = pd.CategoricalDtype(
    categories=[
        '0 a 9 anos','10 a 19 anos', '20 a 29 anos','30 a 39 anos'
        , '40 a 49 anos', '50 a 59 anos', '60 a 69 anos'
        , '70 a 79 anos','80 anos ou mais'
        ]
    ,ordered=True
)

df_pop = pd.read_csv(
    '../../br_demography/results/tab/pop_municipios_rmc_2000_2022_estimativa_intercensitaria.csv'
    , decimal=','
    , sep=';'
    , index_col=['Município', 'Sexo', 'Faixa Etária']
    , dtype={'Faixa Etária':categories}
    )
df_pop.columns = df_pop.columns.astype(int)
df_pop.sort_index(inplace=True)
df_pop.iloc[np.r_[0:2, -2:0]]

df = df_pop[[2022]]
df = df.rename(columns={2022:'População 2022'})
df.iloc[np.r_[0:2, -2:0]]

## Nascimentos

df_birth_rate = pd.read_csv(
    '../../br_demography/results/tab/taxa_natalidade_projetada_rmc_2022_2045_2.csv'
    , decimal=','
    , sep=';'
    , index_col=['Município', 'Faixa Etária']
    , encoding='utf=-16'
    , dtype={'Faixa Etária':categories}
    )
df_birth_rate = df_birth_rate.iloc[:,3:]
df_birth_rate.columns = df_birth_rate.columns.astype(int)
df_birth_rate.sort_index(inplace=True)
df_birth_rate.iloc[np.r_[0:2, -2:0]]

df_nascimentos_2022_2031 = ((
    df_pop.reset_index()
    .loc[df_pop.reset_index()['Sexo'] == 'Feminino'].drop(columns=['Sexo'])
    .set_index(['Município', 'Faixa Etária'])
    .loc[:,2022]
).mul(
    (
    df_birth_rate
    .loc[:,list(range(2022, 2032))]
    .mean(axis=1)
    .to_frame()
    .div(1000)
    .mul(10)
    .squeeze()
    )
)
    .to_frame()
    .reset_index()
    .groupby(by=['Município'])
    .sum()
    .rename(columns={0:'Nascimentos até 2022-2031'})
)

df_nascimentos_2022_2031

df.reset_index()[df.reset_index()['Faixa Etária'] == '0 a 9 anos']

df_nascimentos_2022_2031['Masculino'] = df_nascimentos_2022_2031 * 0.51
df_nascimentos_2022_2031['Feminino'] = df_nascimentos_2022_2031['Nascimentos até 2022-2031'] - df_nascimentos_2022_2031['Masculino']


df['Nascimentos 2022-2031'] = (
    df_nascimentos_2022_2031
    .reset_index()
    .loc[:,['Município', 'Masculino', 'Feminino']]
    .melt(id_vars=['Município'], var_name='Sexo', value_name='Nascimentos 2022-2031')
    .assign(Faixa='0 a 9 anos')
    .rename(columns={'Faixa':'Faixa Etária'})
    .astype({'Faixa Etária': categories})
    .set_index(['Município', 'Sexo', 'Faixa Etária'])
    .astype(int)
    .sort_index()
).squeeze()
df = df.fillna(0)
df = df.astype(int)
df.head(2)

df.loc[df.index.get_level_values('Faixa Etária') == '0 a 9 anos']

## Taxa de Sobrevivência

df_survival_rate = pd.read_csv(
    '../../br_demography/results/tab/taxa_sobrevivencia_projetada_pdui_2022_2045.csv'
    , decimal=','
    , sep=';'
    , index_col=['Município', 'Sexo', 'Faixa Etária']
    , encoding='utf=-16'
    , dtype={'Faixa Etária':categories}
    )
df_survival_rate = df_survival_rate.iloc[:,3:]
df_survival_rate.columns = df_survival_rate.columns.astype(int)
df_survival_rate.sort_index(inplace=True)
df_survival_rate.head(2)

df['Taxa de Sobrevivência 2022-2031'] = df_survival_rate.loc[:,list(range(2022, 2032))].prod(axis=1).to_frame().rename(columns={0:'2032'}).squeeze()

faixas_etarias_anteriores = ['0 a 9 anos', '10 a 19 anos', '20 a 29 anos', '30 a 39 anos', '40 a 49 anos', '50 a 59 anos', '60 a 69 anos', '70 a 79 anos', '80 anos ou mais']
faixas_etarias_posteriores = faixas_etarias_anteriores.copy()

faixas_etarias_posteriores.append(faixas_etarias_anteriores[-1])
faixas_etarias_posteriores = faixas_etarias_posteriores[1:]
dict_faixas_compat = {faixa_anterior:faixa_posterior for faixa_anterior, faixa_posterior in zip(faixas_etarias_anteriores, faixas_etarias_posteriores)}
dict_faixas_compat


df = df.join(
    (
    df
    .assign(Próxima_Faixa_Etária=df.reset_index()['Faixa Etária'].map(dict_faixas_compat).values)
    .reset_index()
    .set_index(['Município', 'Sexo', 'Faixa Etária', 'Próxima_Faixa_Etária'])
    .loc[:,'População 2022']
    .mul(df['Taxa de Sobrevivência 2022-2031'])
    .reset_index()
    .rename(columns={0:'Sobrevivência 2022-2031'})
    .astype({'Sobrevivência 2022-2031':int, 'Próxima_Faixa_Etária':categories})
    .drop(columns=['Faixa Etária'])
    .rename(columns={'Próxima_Faixa_Etária':'Faixa Etária'})
    .groupby(by=['Município', 'Sexo', 'Faixa Etária'])
    .sum()
#    .set_index(['Município', 'Sexo', 'Faixa Etária'])
    )
)

df['População 2032'] = df['Nascimentos 2022-2031'] + df['Sobrevivência 2022-2031']
df_pop[2032] = df['População 2032']

df

# Projeção para 2042

## População do Ano Inicial

## Nascimentos

df['População 2032']

df_nascimentos_2032_2041 = (
    (
    df_pop.reset_index()
    .loc[df_pop.reset_index()['Sexo'] == 'Feminino'].drop(columns=['Sexo'])
    .set_index(['Município', 'Faixa Etária'])
    .loc[:,2032]
).mul(
    (
    df_birth_rate
    .loc[:,list(range(2032, 2042))]
    .mean(axis=1)
    .to_frame()
    .div(1000)
    .mul(10)
    .squeeze()
    )
)
    .to_frame()
    .reset_index()
    .groupby(by=['Município'])
    .sum()
    .rename(columns={0:'Nascimentos 2032-2041'})
)

df_nascimentos_2032_2041['Masculino'] = df_nascimentos_2032_2041 * 0.51

df_nascimentos_2032_2041['Feminino'] = df_nascimentos_2032_2041['Nascimentos 2032-2041'] - df_nascimentos_2032_2041['Masculino']


df['Nascimentos 2032-2041'] = (
    df_nascimentos_2032_2041
    .reset_index()
    .loc[:,['Município', 'Masculino', 'Feminino']]
    .melt(id_vars=['Município'], var_name='Sexo', value_name='Nascimentos 2032-2041')
    .assign(Faixa='0 a 9 anos')
    .rename(columns={'Faixa':'Faixa Etária'})
    .astype({'Faixa Etária': categories})
    .set_index(['Município', 'Sexo', 'Faixa Etária'])
    .astype(int)
    .sort_index()
).squeeze()
df = df.fillna(0)
df = df.astype(int)
df.head(2)

df.loc[df.index.get_level_values('Faixa Etária') == '0 a 9 anos'][['Nascimentos 2022-2031', 'Nascimentos 2032-2041']]

## Taxa de Sobrevivência

df['Taxa de Sobrevivência 2032-2041'] = df_survival_rate.loc[:,list(range(2032, 2042))].prod(axis=1).to_frame().rename(columns={0:'2042'}).squeeze()

faixas_etarias_anteriores = ['0 a 9 anos', '10 a 19 anos', '20 a 29 anos', '30 a 39 anos', '40 a 49 anos', '50 a 59 anos', '60 a 69 anos', '70 a 79 anos', '80 anos ou mais']
faixas_etarias_posteriores = faixas_etarias_anteriores.copy()

faixas_etarias_posteriores.append(faixas_etarias_anteriores[-1])
faixas_etarias_posteriores = faixas_etarias_posteriores[1:]
dict_faixas_compat = {faixa_anterior:faixa_posterior for faixa_anterior, faixa_posterior in zip(faixas_etarias_anteriores, faixas_etarias_posteriores)}
dict_faixas_compat


df = df.join(
    (
    df
    .assign(Próxima_Faixa_Etária=df.reset_index()['Faixa Etária'].map(dict_faixas_compat).values)
    .reset_index()
    .set_index(['Município', 'Sexo', 'Faixa Etária', 'Próxima_Faixa_Etária'])
    .loc[:,'População 2032']
    .mul(df['Taxa de Sobrevivência 2032-2041'])
    .reset_index()
    .rename(columns={0:'Sobrevivência 2032-2041'})
    .astype({'Sobrevivência 2032-2041':int, 'Próxima_Faixa_Etária':categories})
    .drop(columns=['Faixa Etária'])
    .rename(columns={'Próxima_Faixa_Etária':'Faixa Etária'})
    .groupby(by=['Município', 'Sexo', 'Faixa Etária'])
    .sum()
#    .set_index(['Município', 'Sexo', 'Faixa Etária'])
    )
)

df['População 2042'] = df['Nascimentos 2032-2041'] + df['Sobrevivência 2032-2041']

df[['População 2032', 'População 2042']].sum()

df

df.to_csv(path_or_buf='../../br_demography/results/tab/projecao_populacional_rmc_2022_2042_2.csv', sep=';', decimal=',', encoding='utf-16')
